+++ CHANGELOG.md
@@ -41,6 +41,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Users are now steered to create contacts with specific job titles, e.g.
   Headteacher or CEO
+- Change the display of contacts on the external contacts index page
 
 ## [Release-82][release-82]